Override toctree helper to generate valid links
`toctree` helper does not use `pathto` helper to generate the links to
the other documents. So, we need to override the Sphinx's internals
where the toctree is generated and modify the refuri for each node of
the toctree after it's generated.

Our custom function uses the internal TocTree class from Sphinx but
after the toctree is generated it traverses it looking for reference
nodes and replace the refuri.