Resolve internal Sphinx javascript resources URLs
Sphinx adds some common javascript file to all the pages
rendered. These files are rendered using a template tag called
`js_tag` that it's registered inside a an event.

Since the function connected to this event is called after our
extension re-defines the `pathto` function in the context, we need to
re-connect the same function to redefine the `js_tag` but now with our
own patched `pathto` function.