Support Sphinx 4.x
Sphinx 4.x uses `css_tag` in the `basic/layout.html` template now instead of
hardcoding the `pygments.css` and the theme's stylesheet.

We need to use the same method we have been using for JS files: connect a signal
to be called on `html-page-context` for CSS files.

See
https://github.com/sphinx-doc/sphinx/pull/8643/files#diff-a5066e933cbf65adc46e0d1ab9a0b44e0a53ca64cc95dca7e6aa902aed6bd468L98

Note also that the structure of the HTML tag generated changed as well. Tests
were updated to support old and new Sphinx versions.