Read the Docs: handle special URLs
Read the Docs serves their static files under special URLs starting with `/_/`
that should not be modified by this extension. Otherwise, the flyout and other
integrations are broken.

This commit checks if the documentation is building on Read the Docs and if the
URLs starts with `/_/` and skip the manipulation in that case.